chore: no need for diff store and index
